refactor: move trade-executor schema files to schemas dir
