+++ src/lib/trade-executor/schemas/blockchain-transaction.ts
