+++ src/lib/trade-executor/schemas/interest.ts
