+++ src/lib/trade-executor/schemas/loan.ts
