+++ src/lib/trade-executor/schemas/portfolio.ts
