+++ src/lib/trade-executor/schemas/position.ts
@@ -6,14 +6,14 @@
  *
  */
 import { z } from 'zod';
+import { percent, primaryKey, primaryKeyString, unixTimestampToDate, usDollarAmount } from './utility-types';
 import { assetIdentifierSchema, tradingPairIdentifierSchema } from './identifier';
 import { tradeExecutionSchema } from './trade';
-import { percent, primaryKey, primaryKeyString, unixTimestampToDate, usDollarAmount } from './utility-types';
 import { loanSchema } from './loan';
 import { balanceUpdateSchema } from './balance-update';
 import { valuationUpdateSchema } from './valuation';
-import { type TradeInfo, createTradeInfo } from './trade-info';
-import { createTradingPairInfo } from './trading-pair-info';
+import { type TradeInfo, createTradeInfo } from '../state/trade-info';
+import { createTradingPairInfo } from '../state/trading-pair-info';
 
 export const positionStatus = z.enum(['open', 'closed', 'frozen']);
 export type PositionStatus = z.infer<typeof positionStatus>;
