+++ src/lib/trade-executor/schemas/state.ts
