+++ src/lib/trade-executor/schemas/utility-type-fixtures.ts
