+++ src/lib/trade-executor/schemas/utility-types.ts
