+++ src/lib/trade-executor/schemas/valuation.ts
