+++ src/params/positionStatus.ts
@@ -1,4 +1,4 @@
-import { type PositionStatus, positionStatus } from 'trade-executor/state/position';
+import { type PositionStatus, positionStatus } from 'trade-executor/schemas/position';
 
 export function match(param): param is PositionStatus {
 	return positionStatus.safeParse(param).success;
