+++ src/routes/strategies/[strategy].ipynb/+server.ts
@@ -1,6 +1,6 @@
 import { error, text } from '@sveltejs/kit';
 import { publicApiError } from '$lib/helpers/public-api';
-import { configuredStrategies } from 'trade-executor/strategy/configuration';
+import { configuredStrategies } from 'trade-executor/schemas/configuration';
 
 export async function GET({ fetch, params }) {
 	const strategy = configuredStrategies.get(params.strategy);
